Finish adding basic classes, finish generateId util function.

+++ src/oop19_ca2_luke_halpenny/Pet.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
  */
 public class Pet {
 
-    private static int incrementalId = 1;
+    private static long incrementalId = 1;
 
     private String type;
     private String name;
@@ -21,10 +21,10 @@ public class Pet {
     private String colour;
     private Gender gender;
     private LocalDateTime dateRegistered;
-    private int id;
-    private int ownerId;
+    private long id;
+    private long ownerId;
 
-    public Pet(int id, String name, String type, String breed, int age, String colour, Gender gender, int ownerId,
+    public Pet(long id, String name, String type, String breed, int age, String colour, Gender gender, long ownerId,
                LocalDateTime dateRegistered) {
         this.setId(id);
         this.setOwnerId(ownerId);
@@ -43,7 +43,7 @@ public class Pet {
 
     public static Pet newPet(Owner owner, String name, String type, String breed, int age, String colour,
                              Gender gender) {
-        int id = incrementalId;
+        long id = incrementalId;
         ++incrementalId;
         LocalDateTime dateRegistered = LocalDateTime.now();
         return new Pet(id, name, type, breed, age, colour, gender, owner.getId(), dateRegistered);
@@ -108,19 +108,19 @@ public class Pet {
         this.dateRegistered = dateRegistered;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getOwnerId() {
+    public long getOwnerId() {
         return ownerId;
     }
 
-    public void setOwnerId(int ownerId) {
+    public void setOwnerId(long ownerId) {
         this.ownerId = ownerId;
     }
 }
